requirements.txt
Flask==3.1.2
mysql_connector_repackaged==0.3.1
